requirements.txt
paho-mqtt
libscrc
pyserial
bitstring